Docusite_Backend/requirements.txt
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
django-model-utils
django-filter
django-cors-headers